Add -Wdouble-promotion switch to clang/gcc compilers
It introduces a lot of ugly casts. Long term the floats should probably
be changed to doubles everywhere since we're only compiling for 64 bit
x86.

+++ src/tabcontrollers/FixFloorTabController.cpp
@@ -135,15 +135,17 @@ void FixFloorTabController::eventLoopTick(
             {
                 if ( std::abs( tempRoll ) <= M_PI_2 )
                 {
-                    floorOffsetY = tempOffsetY - controllerUpOffsetCorrection;
+                    floorOffsetY = static_cast<float>( tempOffsetY )
+                                   - controllerUpOffsetCorrection;
                 }
                 else
                 {
-                    floorOffsetY = tempOffsetY - controllerDownOffsetCorrection;
+                    floorOffsetY = static_cast<float>( tempOffsetY )
+                                   - controllerDownOffsetCorrection;
                 }
 
-                floorOffsetX = tempOffsetX;
-                floorOffsetZ = tempOffsetZ;
+                floorOffsetX = static_cast<float>( tempOffsetX );
+                floorOffsetZ = static_cast<float>( tempOffsetZ );
 
                 LOG( INFO )
                     << "Fix Floor and adjust room center: Floor Offset = ["
